#3710 directly removing old endpoint and switching to new
